Consolidate tests into one file
